rename abstract dir to model
