+++ app/src/main/java/com/slothnull/android/medox/Settings.java
@@ -21,7 +21,7 @@ import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
-import com.slothnull.android.medox.Abstract.AbstractConfig;
+import com.slothnull.android.medox.model.AbstractConfig;
 
 public class Settings extends AppCompatActivity implements LocationListener {
