+++ app/src/main/java/com/slothnull/android/medox/fcm/MyFirebaseMessagingService.java
@@ -40,7 +40,7 @@ import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.messaging.FirebaseMessagingService;
 import com.google.firebase.messaging.RemoteMessage;
-import com.slothnull.android.medox.Abstract.AbstractNotification;
+import com.slothnull.android.medox.model.AbstractNotification;
 import com.slothnull.android.medox.Home;
 import com.slothnull.android.medox.R;
 import com.slothnull.android.medox.SeniorHome;
