+++ app/src/main/java/com/slothnull/android/medox/fragment/NotificationFragment.java
@@ -17,7 +17,7 @@ import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.Query;
-import com.slothnull.android.medox.Abstract.AbstractNotification;
+import com.slothnull.android.medox.model.AbstractNotification;
 import com.slothnull.android.medox.R;
 import com.slothnull.android.medox.viewholder.NotificationViewHolder;
