+++ app/src/main/java/com/slothnull/android/medox/fragment/ScheduleFragment.java
@@ -24,8 +24,8 @@ import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.Query;
 import com.google.firebase.database.ValueEventListener;
-import com.slothnull.android.medox.Abstract.AbstractConfig;
-import com.slothnull.android.medox.Abstract.AbstractSchedule;
+import com.slothnull.android.medox.model.AbstractConfig;
+import com.slothnull.android.medox.model.AbstractSchedule;
 import com.slothnull.android.medox.AddSchedule;
 import com.slothnull.android.medox.R;
 import com.slothnull.android.medox.viewholder.ScheduleViewHolder;
