+++ app/src/main/java/com/slothnull/android/medox/fragment/SeniorEmergencyFragment.java
@@ -11,7 +11,6 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.graphics.BitmapFactory;
 import android.media.RingtoneManager;
-import android.net.ConnectivityManager;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
@@ -32,11 +31,11 @@ import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
-import com.slothnull.android.medox.Abstract.AbstractCommand;
-import com.slothnull.android.medox.Abstract.AbstractConfig;
-import com.slothnull.android.medox.Abstract.AbstractEmergency;
-import com.slothnull.android.medox.Abstract.AbstractMobileToken;
-import com.slothnull.android.medox.Abstract.AbstractWarehouse;
+import com.slothnull.android.medox.model.AbstractCommand;
+import com.slothnull.android.medox.model.AbstractConfig;
+import com.slothnull.android.medox.model.AbstractEmergency;
+import com.slothnull.android.medox.model.AbstractMobileToken;
+import com.slothnull.android.medox.model.AbstractWarehouse;
 import com.slothnull.android.medox.R;
 import com.slothnull.android.medox.SeniorHome;
