+++ app/src/main/java/com/slothnull/android/medox/fragment/WarehouseFragment.java
@@ -21,10 +21,10 @@ import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
-import com.slothnull.android.medox.Abstract.AbstractCommand;
-import com.slothnull.android.medox.Abstract.AbstractConfig;
-import com.slothnull.android.medox.Abstract.AbstractData;
-import com.slothnull.android.medox.Abstract.AbstractWarehouse;
+import com.slothnull.android.medox.model.AbstractCommand;
+import com.slothnull.android.medox.model.AbstractConfig;
+import com.slothnull.android.medox.model.AbstractData;
+import com.slothnull.android.medox.model.AbstractWarehouse;
 import com.slothnull.android.medox.R;
 
 import java.util.ArrayList;
