+++ app/src/main/java/com/slothnull/android/medox/model/AbstractMobileToken.java
@@ -1,25 +1,25 @@
-package com.slothnull.android.medox.Abstract;
+package com.slothnull.android.medox.model;
 
 /**
  * Created by Shaheen on 17-Mar-17
  * Project: seniormedox
- * Package: com.slothnull.android.medox
+ * Package: com.slothnull.android.seniormedox
  */
 
 import com.google.firebase.database.IgnoreExtraProperties;
 
 // [START user_class]
 @IgnoreExtraProperties
-public class AbstractWatchToken {
+public class AbstractMobileToken {
 
-    public String watch;
+    public String mobile;
 
-    public AbstractWatchToken() {
+    public AbstractMobileToken() {
         // Default constructor required for calls to DataSnapshot.getValue(User.class)
     }
 
-    public AbstractWatchToken(String token) {
-        this.watch = token;
+    public AbstractMobileToken(String token) {
+        this.mobile = token;
     }
 
 }
