+++ app/src/main/java/com/slothnull/android/medox/model/AbstractNotification.java
@@ -1,4 +1,4 @@
-package com.slothnull.android.medox.Abstract;
+package com.slothnull.android.medox.model;
 
 /**
  * Created by Shaheen on 11-Mar-17
